Dockerfile: Add
The Dockerfile for building Yocto must be located in the same repository
as the Jenkinsfile. This is caused by Jenkins using the Dockerfile as a
stage agent and there is no possibility to obtain the Dockerfile from
some stash before building the Docker image.

This Dockerfile equals the Dockerfile currently used in cml kirkstone.

Signed-off-by: Johannes Wiesboeck <johannes.wiesboeck@aisec.fraunhofer.de>