gyroidos-base: Do not checkout cml
Checking out the cml repository and building with externalsrc repeatedly
leads to the problem that, if we add a patch for the cml in a project
yocto layer, this patch is not applied due to externalsrc. Therefore,
bugs or changes go unnoticed during testing.

Let's remove this feature once and forall. IF someone want's to apply
changes to the cml, the should use Yocto's 'devtool modify' to do so:

devtool modify cmld

The cml source can then be found at
out-yocto/tmp/workspace/sources/cmld from where it is also built using
externalsrc (no commits required) and with all patches from layers
already applied.

Note that changes to service(-static) need to be done using

devtool modify service(-static)

Signed-off-by: Johannes Wiesböck <johannes.wiesboeck@aisec.fraunhofer.de>